requirements.txt
pandas==1.3.5
python-dotenv==0.19.2
python_binance==1.0.15
tweepy==4.4.0
